refactor: change files' directory
