+++ src/main/java/org/standard/dreamcalendar/config/mail/AuthCodeGenerator.java
@@ -1,4 +1,4 @@
-package org.standard.dreamcalendar.config.auth;
+package org.standard.dreamcalendar.config.mail;
 
 import org.springframework.stereotype.Component;
