+++ src/main/java/org/standard/dreamcalendar/domain/email/EmailAuthService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.standard.dreamcalendar.config.auth.AuthCodeGenerator;
+import org.standard.dreamcalendar.config.mail.AuthCodeGenerator;
 import org.standard.dreamcalendar.service.DtoConverter;
 
 @Service
